Added missing Recursive constructor, python implementation with refactor
